fix(pdf): address inference server review feedback
